requirements.txt
streamlit==1.31.0
pandas==1.5.3
requests==2.31.0
plotly==5.18.0
pillow==10.1.0
numpy==1.25.2
python-dotenv==1.0.0